duplicate/requirements.txt
dnspython==2.7.0
greenlet==3.1.1
numpy==2.2.3
opencv-python==4.11.0.86
packaging==24.2
pika==1.3.2
pillow==11.1.0
psycopg2-binary==2.9.10
pymongo==4.11.1
pytesseract==0.3.13
SQLAlchemy==2.0.38
typing_extensions==4.12.2
